Separate out refs::* visitor to a child trait
