+++ graphql-parser/src/query/visit.rs
@@ -1,20 +1,14 @@
 use super::{Definition, Document, Selection, SelectionSet};
-use crate::query::refs::{SelectionRef, SelectionSetRef};
-use crate::{visit, visit_each};
+use crate::{node_trait, visit, visit_each};
 
-// TODO(ran)(p2) FIXME: consider separating the visitor to RefVisitor
 #[allow(unused_variables)]
 pub trait Visitor {
     fn enter_query(&mut self, doc: &Document) {}
     fn enter_query_def(&mut self, def: &Definition) {}
     fn enter_sel_set(&mut self, sel_set: &SelectionSet) {}
-    fn enter_sel_set_ref(&mut self, sel_set: &SelectionSetRef) {}
     fn enter_sel(&mut self, sel: &Selection) {}
-    fn enter_sel_ref(&mut self, sel: &SelectionRef) {}
     fn leave_sel(&mut self, sel: &Selection) {}
-    fn leave_sel_ref(&mut self, sel: &SelectionRef) {}
     fn leave_sel_set(&mut self, sel_set: &SelectionSet) {}
-    fn leave_sel_set_ref(&mut self, sel_set: &SelectionSetRef) {}
     fn leave_query_def(&mut self, def: &Definition) {}
     fn leave_query(&mut self, doc: &Document) {}
 }
@@ -23,9 +17,7 @@ pub trait Map: visit::Map {
     fn query(&mut self, doc: &Document, stack: &[Self::Output]) -> Self::Output;
     fn query_def(&mut self, def: &Definition, stack: &[Self::Output]) -> Self::Output;
     fn sel_set(&mut self, sel_set: &SelectionSet, stack: &[Self::Output]) -> Self::Output;
-    fn sel_set_ref(&mut self, sel_set: &SelectionSetRef, stack: &[Self::Output]) -> Self::Output;
     fn sel(&mut self, sel: &Selection, stack: &[Self::Output]) -> Self::Output;
-    fn sel_ref(&mut self, sel: &SelectionRef, stack: &[Self::Output]) -> Self::Output;
 }
 
 impl<M: Map> Visitor for visit::Fold<M> {
@@ -38,35 +30,15 @@ impl<M: Map> Visitor for visit::Fold<M> {
     fn enter_sel_set(&mut self, sel_set: &SelectionSet) {
         self.stack.push(self.map.sel_set(sel_set, &self.stack));
     }
-
-    fn enter_sel_set_ref(&mut self, sel_set: &SelectionSetRef) {
-        self.stack.push(self.map.sel_set_ref(sel_set, &self.stack));
-    }
-
     fn enter_sel(&mut self, sel: &Selection) {
         self.stack.push(self.map.sel(&sel, &self.stack));
     }
-
-    fn enter_sel_ref(&mut self, sel: &SelectionRef) {
-        self.stack.push(self.map.sel_ref(&sel, &self.stack));
-    }
-
     fn leave_sel(&mut self, _sel: &Selection) {
         self.pop();
     }
-
-    fn leave_sel_ref(&mut self, _sel: &SelectionRef) {
-        self.pop();
-    }
-
     fn leave_sel_set(&mut self, _sel_set: &SelectionSet) {
         self.pop();
     }
-
-    fn leave_sel_set_ref(&mut self, _sel_set: &SelectionSetRef) {
-        self.pop();
-    }
-
     fn leave_query_def(&mut self, _def: &Definition) {
         self.pop();
     }
@@ -75,18 +47,7 @@ impl<M: Map> Visitor for visit::Fold<M> {
     }
 }
 
-pub trait Node {
-    fn accept<V: Visitor>(&self, visitor: &mut V);
-    fn map<M: Map>(&self, map: M) -> visit::Fold<M> {
-        let mut mapping = visit::Fold {
-            stack: vec![],
-            map,
-            output: None,
-        };
-        self.accept(&mut mapping);
-        mapping
-    }
-}
+node_trait!(Visitor, Map);
 
 impl<'a> Node for Document<'a> {
     fn accept<V: Visitor>(&self, visitor: &mut V) {
@@ -132,7 +93,6 @@ impl<'a> Node for Selection<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::query::refs::{SelectionRef, SelectionSetRef};
     use crate::{parse_query, query, query::*, visit, ParseError};
 
     #[test]
@@ -257,20 +217,8 @@ mod tests {
             ) -> Self::Output {
                 format!("{}sel_set", "    ".repeat(stack.len()))
             }
-            fn sel_set_ref(
-                &mut self,
-                _sel_set: &SelectionSetRef,
-                _stack: &[Self::Output],
-            ) -> Self::Output {
-                unreachable!()
-            }
-
             fn sel<'a>(&mut self, _: &Selection<'a>, stack: &[Self::Output]) -> Self::Output {
                 format!("{}sel", "    ".repeat(stack.len()))
-            }
-
-            fn sel_ref(&mut self, _sel: &SelectionRef, _stack: &[Self::Output]) -> Self::Output {
-                unreachable!()
             }
         }
