+++ query-planner/src/context.rs
@@ -3,7 +3,8 @@ use crate::consts::{typename_field_def, typename_field_node};
 use crate::federation::Federation;
 use crate::helpers::Op;
 use crate::visitors::VariableUsagesMap;
-use graphql_parser::query::refs::{FieldRef, SelectionRef, SelectionSetRef};
+use graphql_parser::query::refs::{FieldRef, Node, SelectionRef, SelectionSetRef};
+use graphql_parser::query::Node as QueryNode;
 use graphql_parser::query::*;
 use graphql_parser::schema::TypeDefinition;
 use graphql_parser::{schema, Name};
